implémentation du test de parsing Json
# Conflicts:
#	src/main/java/fr/insalyon/pldagile/sortie/FeuilleDeRoute.java
#	target/classes/static/index.html

+++ src/main/java/fr/insalyon/pldagile/sortie/parseurTourneeJson.java
@@ -39,8 +39,8 @@ public class parseurTourneeJson {
                 JSONObject noeudDePassageArriveeJson = cheminObject.getJSONObject("NoeudDePassageArrivee");
                 for (int j = 0; j < tronconArray.length(); j++) {
                     JSONObject tronconObject = tronconArray.getJSONObject(j);
-                    Troncon troncon = new Troncon(tronconObject.getInt("idOrigine"),
-                            tronconObject.getInt("idDestination"),
+                    Troncon troncon = new Troncon(tronconObject.getLong("idOrigine"),
+                            tronconObject.getLong("idDestination"),
                             tronconObject.getDouble("longueur"),
                             tronconObject.getString("nomRue"));
                     listTroncons.add(troncon);
